worker/requirements.txt
affine==2.4.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dask==2025.11.0
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.61.0
fsspec==2025.10.0
geocube==0.7.1
geopandas==1.1.1
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kafka-python==2.3.0
kiwisolver==1.4.9
locket==1.0.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.3.5
odc-geo==0.4.10
packaging==25.0
pandas==2.3.3
parso==0.8.5
partd==1.4.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
pystac==1.14.1
pystac-client==0.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.4.3
referencing==0.37.0
requests==2.32.5
rioxarray==0.20.0
rpds-py==0.30.0
scipy==1.16.3
seaborn==0.13.2
shapely==2.1.2
six==1.17.0
squarify==0.4.4
stack-data==0.6.3
stackstac==0.5.1
toolz==1.1.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xarray==2025.11.0
